+++ app/Routes.js
@@ -17,7 +17,7 @@ import SupportSettingsPage from './containers/settings/categories/SupportSetting
 import TermsOfUseSettingsPage from './containers/settings/categories/TermsOfUseSettingsPage';
 import TermsOfUsePage from './containers/profile/TermsOfUsePage';
 import WalletSettingsPage from './containers/settings/categories/WalletSettingsPage';
-import AboutYoroiPage from './components/settings/categories/AboutYoroiPage';
+import AboutYoroiSettings from './components/settings/categories/AboutYoroiSettings';
 
 // Dynamic container loading - resolver loads file relative to '/app/' directory
 const LoadingPage = resolver('containers/LoadingPage');
@@ -131,7 +131,7 @@ const SettingsSubpages = (stores, actions) => (
     <Route
       exact
       path={ROUTES.SETTINGS.ABOUT_YOROI}
-      component={(props) => <AboutYoroiPage {...props} stores={stores} actions={actions} />}
+      component={(props) => <AboutYoroiSettings {...props} stores={stores} actions={actions} />}
     />
     <Redirect to={ROUTES.SETTINGS.GENERAL} />
   </Switch>
