--- app/components/layout/GridFlexContainer.scss
@@ -1,7 +1,0 @@
-.component {
-  width: 100%;
-  display: flex;
-  flex-direction: row;
-  flex-wrap: wrap;
-  position: relative;
-}
