+++ app/components/layout/HorizontalFlexContainer.scss
@@ -1,6 +1,5 @@
 .component {
   width: 100%;
-  height: 100%;
   display: flex;
   flex-direction: row;
   overflow-x: overlay;
