+++ app/components/settings/categories/AboutYoroiSettings.scss
@@ -1,0 +1,12 @@
+.component {}
+
+.yoroiLogo {
+  & > svg {
+    display: block;
+    width: 100px;
+    margin: 0 114px;
+    & > path {
+      fill: var(--theme-button-primary-background-color);
+    }
+  }
+}
